// vintage-maroc-front/src/app/models/vinyl.model.ts
export interface VinylRecord {
  id: string;
  title: string;
  artist: string;
  imageUrl: string;
  have: number;
  want: number;
  rating: string;
  label: string;
  catalogNumber: string;
  mediaCondition: string;
  sleeveCondition: string;
  description: string;
  ratingPercent: number;
  ratings: number;
  shipsFrom: string;
  price: number;
  shipping: number;
  totalEuros: number;
} 